Refactor useEndpointData
