+++ ee/app/engagement-dashboard/client/components/UsersTab/UsersByTimeOfTheDaySection.js
@@ -4,8 +4,8 @@ import moment from 'moment';
 import React, { useMemo, useState } from 'react';
 
 import { useTranslation } from '../../../../../../client/contexts/TranslationContext';
+import { useEndpointData } from '../../../../../../client/hooks/useEndpointData';
 import { Section } from '../Section';
-import { useEndpointData } from '../../hooks/useEndpointData';
 
 export function UsersByTimeOfTheDaySection() {
 	const t = useTranslation();
@@ -47,7 +47,7 @@ export function UsersByTimeOfTheDaySection() {
 		end: period.end.toISOString(),
 	}), [period]);
 
-	const data = useEndpointData('GET', 'engagement-dashboard/users/users-by-time-of-the-day-in-a-week', params);
+	const data = useEndpointData('engagement-dashboard/users/users-by-time-of-the-day-in-a-week', params);
 
 	const [
 		dates,
